requirements.txt
click
python-dotenv
langchain
langchain-community
langchain-cohere
langchain-experimental
langchain-groq
langchain-pinecone
langchain-fireworks
langchain_google_genai
langchain-groq
langchain_mistralai
langchain-openai
langchain-openrouter
langchain-ollama
langchain-together
langchain-anthropic
langchain-aws
pdfplumber
pyperclip
rich
selenium
spacy
scikit-learn
trafilatura
webdriver-manager
